Parse Uemis cylinder data correctly.
There's a big comment there now about what is going on.  It took me a
while to understand how the crazy seven-tank uemis dive computer
information actually works.

So the Uemis computer has 4 different "tank profiles":
 - single tank air (0)
 - single tank nitrox (1)
 - two-tank nitrox (2)
 - three-tank nitrox (3)
and the computer always lists all seven tank cases (because that's how
you fill them in).

Depending on the "gas.template" you are supposed to then *use* just one
particular profile.  Why the computer doesn't just give you the tanks
for that one profile, who knows? It seems to be more of the same "Uemis
dive data isn't so much about the dive, it's about dive computer state"
mentality.

So we first get the profile information, and then based on that we need
to pick the right tanks from the set of seven that we're presented with.

All clear?

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>