Now that we don't mess up import, we can save the cylinder working pressure
We used to have the heuristic that if we saw a cylinder working
pressure, then the cylinder size would be in cuft.  Which meant that we
couldn't export our working pressures, because it would mess things up
on import.

But working pressure is actually nice to know, if you ever work with
cylinders in imperial units.  So now that the import is fixed, add the
working pressure to the export.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>