Remove gtk references from profile.c - but keep using cairo
Cairo has many backends and is a very nice drawing framework to use here.

We split out the gtk related declarations from display.h into
display-gtk.h. For now we need to add this include file to the other .c
files where we haven't separated out the gtk dependent parts, but that's
of course intended to be only temporary.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>