Don't mark the divelist window transient
That also makes it always stay in front of the other window, which is
just annoying.  I only did it because I wanted to make sure it dies when
the main window does, but since we just kill the main loop when closing
either window, that just isn't an issue.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>