Add note on dive computers using the same import engine
Lots of dive computers are just variations on a theme, or sometimes even
just rebadged copies of each others with different manufacturer and
model names.  The import dialog may not mention your exact dive computer
by name, but that doesn't necessarily mean that you cannot import data
from it.

Make that clearer in the README, and list the rough list of dive
computers supported by libdivecomputer.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>