We forgot to pick up the 'value' field of a dive event
Just missed that one entirely in the xml parser for some reason.
Probably because the fields don't have much semantic meaning, so I
didn't even realize that I had missed one of the random integer values
in an event.

On my suunto, the 'value' field seems to contain things like the new
Oxygen percentage of a gas change event etc.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>